Extract DescriptionEditor component to common folder and replace Stack with Box in PriceSection
