+++ frontend/src/components/AddPage/AddPage.styles.tsx
@@ -97,39 +97,3 @@ export const SubmitButtonStyle: SxProps<Theme> = {
         backgroundColor: 'rgba(236, 142, 6, 1)',
     },
 };
-
-
-export const DescriptionWrapperStyle: SxProps<Theme> = {
-    position: 'relative',
-    width: '100%',
-    border: '1px solid',
-    borderColor: 'grey.400',
-    borderRadius: 1,
-    px: 1,
-    pt: 1,
-};
-
-export const DescriptionLabelStyle: SxProps<Theme> = {
-    position: 'absolute',
-    top: '-10px',
-    left: '12px',
-    backgroundColor: 'white',
-    px: 0.5,
-    fontSize: '0.75rem',
-    color: 'text.secondary',
-};
-
-export const EditorContentStyle = {
-    '& .ProseMirror': {
-        minHeight: '5em',
-        padding: '4px',
-        overflowY: 'auto',
-    },
-};
-
-export const EditorToolbarStyle: SxProps<Theme> = {
-    padding: '2px 0 0 0',
-    '& svg': {
-        fontSize: '18px',
-    },
-};
